refactor(workstation): extract runWorkflowAction into useWorkflowAction (0.72 app.ts decomposition)
